image2text_app/requirements.txt
Django>=3.2,<4.0  # Django framework
djangorestframework>=3.12  # For RESTful API if needed
Pillow>=8.0  # For image processing
google-cloud-vision>=2.0  # For Google Cloud Vision API
openai>=0.10  # For OpenAI GPT models
mysqlclient>=2.0  # MySQL database adapter
django-environ>=0.4  # For environment variable management
whitenoise>=5.0  # For serving static files in production
celery>=5.0  # For asynchronous task queues
redis>=3.5  # As a broker for Celery
drf-yasg>=1.20  # For API documentation
pytest-django>=4.0
pytest-cov>=2.10
django-debug-toolbar>=3.2

